Trying out embedded frameworks for Skia on iOS
 - Created a libskia_ios xcode project to be the embedded framework
 - Changed the binding to use the new embedded framework

The theory is that we can rely on the xcode linker to strip out the
symbols that are not referenced in the c api.  If a new c file is added
to skia core we will have to change libskia_ios/SkiaKeeper.c to reference
one of the methods in that file.

From what I can tell it works pretty well.  Now the iOS binding is only 7MB
and that is all that is added to the size of the release iphone app size.