Add a Xamarin Workbooks integration
The integration uses the new Xamarin Workbooks SDK and provides some
convenience renderers for SKImage, SKBitmap, SKSurface, SKPixmap, and SKColor to
aid in using those types in a workbook. With the integration loaded, those Skia
objects will be rendered directly in the viewport as images (in the case of the
first 4), or as a Workbooks color in the case of SKColor.