Update docs to reflect SKImage.Encode() changes
It appears that `SKImage.Encode()` now returns an `SKData` object, not a `byte[]` array, so the existing example that used `File.WriteAllBytes()` no longer compiles.  I have updated the example to use `File.OpenWrite()` and the `SKData.SaveTo()` method, which seems to be the closest reasonable parallel example (and, importantly, which actually compiles and works).