Assemblies are required in platform ref folders (#1044)
All the assemblies are required to be in the ref folder as if there is any ref folder, it will take precedence over the lib folder. In the case where we have JUST the netstandard library in the ref, that is used for ALL platforms. Even though the correct assembly is swapped out at compile time, we actually have more things inside the assembly for the platforms.