Correctly decode images (#1055)
When decoding from SKData, it is not immediate. Rather, the data is passed to
a lazy image that is decoded when the pixels are actually needed. Thus,
passing fixed the byte[] array will result in invalid images if the array
is moved in memory or changed later on. So, we create a copy.

The other cases where the codec was use to construct a bitmap first are also
wrong. SKBitmap constructs differently to SKImage, and since SKImage is just
going to create a copy anyway, just copy the encoded data once, and pass the
data to SKImage. This way, we are all in with SKImage, and we get the expected
results.