Dispose the native object first
This is to ensure that when the next call to create a managed object that has the same native handle (such as in a multithreaded environment) the managed object is always brand new. And, no matter what happens to the native object - such as a disposal/recreate or a swap - the manage object is unaffected.
Fixes #1343