SKTypeface does not count bytes but glyphs (#1399)
Historically, SkTypeface.charsToGlyphs assumed the data was already in glyphs. So a UTF32 data array would pass the number of glyphs. Since that method was removed, we switched to SkFont.textToGlyphs. This worked in bytes. As a result, we need to make sure we convert from glyph count to byte count when we use it.