Disable the unused and problematic tools (#2873)
* Disable the unused and problematic tools

The Go modules are all far out of date and some tooling is reporting the
repo as having security issues. We do not use any part of Go, EMSDK or
perfetto in the SkiaSharp project.

In an ideal world we don't have to make changes and there is a special
flag that exists for us to "disable all issues", but that does not exist
and trying to make that flag work only on CI is unnecessary. We can just
delete the few places that are having issues.