Add more SKPicture APIs (#2883)
Adding a few new APIs to help with determining picture and drawable "size" 
as well as adding the `Playback` method to `SKPicture`.

The playback method is different to draw in that draw results in a single
"draw the picture" command, but playback effectively replicates what
was done to create the picture. There is a small different to the latter
in that the picture may optimize/combine commands as well as split 
commands to best fit into the picture's storage.