Log more information about disk space and skip the MSVC builds on public (#3090)
* Try and get more information as to disk space
* Only build MSVC on the internal build

The MSVC compiler is so heavy that the outputs are over 10.8GB
whereas the LLVM compiler is just under 10.1GB.

The public agent only starts with 12GB and after the checkout we are
left with 10.67GB which is smaller than the 10.8GB required. The LLVM
compiler just manages to pass under the bar with about 600MB spare.

If the SkiaSharp artifacts get any bigger, then we will need custom
machines for the public build.