Avoid exception and console error message when skia canvas is created and then immediately disposed.
In javascript methods, return earlier when element is not available in DOM, also remove the error message in same scenario.

In Balazor, the elements are removed from DOM before calling Dispose() of the component. So after each async method, the DOM element may be removed due to navigate to other node. No simple way to detect if the missing element is because the component is disposed, or software logic error.

Ref: https://github.com/dotnet/aspnetcore/issues/45777